pyproject.toml
[tool.poetry]
name = "dcftest"
version = "0.1.0"
description = ""
authors = ["benjaminderei <stygmate@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^2.2"
psycopg2-binary = ">=2.8,<2.9"
django-computedfields = "^0.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
